pyproject.toml
[project]
name = "ai-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "langchain>=1.2.15",
    "langchain-ollama>=1.1.0",
    "langchain-openai>=1.2.0",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]
